requirements.txt
cryptography==41.0.7
PySide6==6.9.0
argon2-cffi==23.1.0 